requirements.txt
blinker==1.9.0
click==8.3.1
colorama==0.4.6
Flask==3.1.3
fonttools==4.61.1
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==6.0.2
MarkupSafe==3.0.3
numpy==2.2.6
opencv-python-headless==4.13.0.92
pdf2docx==0.5.9
PyMuPDF==1.27.1
python-docx==1.2.0
reportlab==4.2.5
termcolor==3.3.0
typing_extensions==4.15.0
Werkzeug==3.1.5
